Lift form state to WorkflowDetail
