+++ public/pages/workflow_detail/resizable_workspace.tsx
@@ -3,9 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useRef, useState, useEffect } from 'react';
-import { Form, Formik } from 'formik';
-import * as yup from 'yup';
+import React, { useRef, useState } from 'react';
 import {
   EuiCodeBlock,
   EuiEmptyPrompt,
@@ -14,20 +12,8 @@ import {
   EuiResizableContainer,
   EuiText,
 } from '@elastic/eui';
-
-import {
-  Workflow,
-  WorkflowConfig,
-  WorkflowFormValues,
-  WorkflowSchema,
-  customStringify,
-} from '../../../common';
-import {
-  isValidUiWorkflow,
-  reduceToTemplate,
-  uiConfigToFormik,
-  uiConfigToSchema,
-} from '../../utils';
+import { Workflow, WorkflowConfig, customStringify } from '../../../common';
+import { reduceToTemplate } from '../../utils';
 import { WorkflowInputs } from './workflow_inputs';
 import { Workspace } from './workspace';
 import { Tools } from './tools';
@@ -37,7 +23,14 @@ import './workspace/workspace-styles.scss';
 import '../../global-styles.scss';
 
 interface ResizableWorkspaceProps {
-  workflow?: Workflow;
+  workflow: Workflow | undefined;
+  isValidWorkflow: boolean;
+  uiConfig: WorkflowConfig | undefined;
+  setUiConfig: (uiConfig: WorkflowConfig) => void;
+  ingestDocs: string;
+  setIngestDocs: (docs: string) => void;
+  query: string;
+  setQuery: (query: string) => void;
 }
 
 const WORKFLOW_INPUTS_PANEL_ID = 'workflow_inputs_panel_id';
@@ -49,27 +42,6 @@ const TOOLS_PANEL_ID = 'tools_panel_id';
  * panels - the ReactFlow workspace panel and the selected component details panel.
  */
 export function ResizableWorkspace(props: ResizableWorkspaceProps) {
-  // Workflow state
-  const [workflow, setWorkflow] = useState<Workflow | undefined>(
-    props.workflow
-  );
-
-  // Formik form state
-  const [formValues, setFormValues] = useState<WorkflowFormValues>({});
-  const [formSchema, setFormSchema] = useState<WorkflowSchema>(yup.object({}));
-
-  // ingest state
-  const [ingestDocs, setIngestDocs] = useState<string>('');
-
-  // query state
-  const [query, setQuery] = useState<string>('');
-
-  // Temp UI config state. For persisting changes to the UI config that may
-  // not be saved in the backend (e.g., adding / removing an ingest processor)
-  const [uiConfig, setUiConfig] = useState<WorkflowConfig | undefined>(
-    undefined
-  );
-
   // Preview side panel state. This panel encapsulates the tools panel as a child resizable panel.
   const [isPreviewPanelOpen, setIsPreviewPanelOpen] = useState<boolean>(true);
   const collapseFnHorizontal = useRef(
@@ -82,12 +54,6 @@ export function ResizableWorkspace(props: ResizableWorkspaceProps) {
     setIsPreviewPanelOpen(!isPreviewPanelOpen);
   };
 
-  // ingest state
-  const [ingestResponse, setIngestResponse] = useState<string>('');
-
-  // query state
-  const [queryResponse, setQueryResponse] = useState<string>('');
-
   // Tools side panel state
   const [isToolsPanelOpen, setIsToolsPanelOpen] = useState<boolean>(true);
   const collapseFnVertical = useRef(
@@ -98,173 +64,125 @@ export function ResizableWorkspace(props: ResizableWorkspaceProps) {
     setIsToolsPanelOpen(!isToolsPanelOpen);
   };
 
-  // workflow state
-  const [isValidWorkflow, setIsValidWorkflow] = useState<boolean>(true);
+  // ingest / search response states to be populated in the Tools panel
+  const [ingestResponse, setIngestResponse] = useState<string>('');
+  const [queryResponse, setQueryResponse] = useState<string>('');
 
-  // Hook to check if the workflow is valid or not
-  useEffect(() => {
-    const missingUiFlow = props.workflow && !isValidUiWorkflow(props.workflow);
-    if (missingUiFlow) {
-      setIsValidWorkflow(false);
-    } else {
-      setWorkflow(props.workflow);
-    }
-  }, [props.workflow]);
-
-  // Initialize the form state based on the workflow's config, if applicable.
-  useEffect(() => {
-    if (workflow?.ui_metadata?.config) {
-      setUiConfig(workflow.ui_metadata.config);
-    }
-  }, [workflow]);
-
-  // Initialize the form state based on the current UI config
-  useEffect(() => {
-    if (uiConfig) {
-      const initFormValues = uiConfigToFormik(uiConfig, ingestDocs);
-      const initFormSchema = uiConfigToSchema(uiConfig);
-      setFormValues(initFormValues);
-      setFormSchema(initFormSchema);
-    }
-  }, [uiConfig]);
-
-  return isValidWorkflow ? (
-    <Formik
-      enableReinitialize={true}
-      initialValues={formValues}
-      validationSchema={formSchema}
-      onSubmit={(values) => {}}
-      validate={(values) => {}}
+  return props.isValidWorkflow ? (
+    <EuiResizableContainer
+      direction="horizontal"
+      className="stretch-absolute"
+      style={{
+        marginLeft: '-8px',
+        marginTop: '-8px',
+      }}
     >
-      {(formikProps) => (
-        <Form>
-          <EuiResizableContainer
-            direction="horizontal"
-            className="stretch-absolute"
-            style={{
-              marginLeft: '-8px',
-              marginTop: '-8px',
-            }}
-          >
-            {(EuiResizablePanel, EuiResizableButton, { togglePanel }) => {
-              if (togglePanel) {
-                collapseFnHorizontal.current = (
-                  panelId: string,
-                  { direction }
-                ) => togglePanel(panelId, { direction });
-              }
-
-              return (
-                <>
-                  <EuiResizablePanel
-                    id={WORKFLOW_INPUTS_PANEL_ID}
-                    mode="main"
-                    initialSize={60}
-                    minSize="25%"
-                    paddingSize="s"
-                  >
-                    <WorkflowInputs
-                      workflow={props.workflow}
-                      uiConfig={uiConfig}
-                      setUiConfig={setUiConfig}
-                      setIngestResponse={setIngestResponse}
-                      setQueryResponse={setQueryResponse}
-                      ingestDocs={ingestDocs}
-                      setIngestDocs={setIngestDocs}
-                      query={query}
-                      setQuery={setQuery}
-                    />
-                  </EuiResizablePanel>
-                  <EuiResizableButton />
-                  <EuiResizablePanel
-                    id={PREVIEW_PANEL_ID}
-                    style={{
-                      marginRight: isPreviewPanelOpen ? '-32px' : '0px',
-                      marginBottom: isToolsPanelOpen ? '0px' : '24px',
-                    }}
-                    mode="collapsible"
-                    initialSize={60}
-                    minSize="25%"
-                    paddingSize="s"
-                    onToggleCollapsedInternal={() => onTogglePreviewChange()}
-                  >
-                    <EuiResizableContainer
-                      className="workspace-panel"
-                      direction="vertical"
-                      style={{
-                        marginLeft: '-8px',
-                        marginTop: '-8px',
-                        padding: 'none',
-                      }}
-                    >
-                      {(
-                        EuiResizablePanel,
-                        EuiResizableButton,
-                        { togglePanel }
-                      ) => {
-                        if (togglePanel) {
-                          collapseFnVertical.current = (
-                            panelId: string,
-                            { direction }
-                          ) =>
-                            // ignore is added since docs are incorrectly missing "top" and "bottom"
-                            // as valid direction options for vertically-configured resizable panels.
-                            // @ts-ignore
-                            togglePanel(panelId, { direction });
-                        }
-
-                        return (
-                          <>
-                            <EuiResizablePanel
-                              mode="main"
-                              initialSize={60}
-                              minSize="25%"
-                              paddingSize="s"
-                              style={{ marginBottom: '-8px' }}
-                            >
-                              <EuiFlexGroup
-                                direction="column"
-                                gutterSize="s"
-                                style={{ height: '100%' }}
-                              >
-                                <EuiFlexItem>
-                                  <Workspace
-                                    workflow={props.workflow}
-                                    uiConfig={uiConfig}
-                                  />
-                                </EuiFlexItem>
-                              </EuiFlexGroup>
-                            </EuiResizablePanel>
-                            <EuiResizableButton />
-                            <EuiResizablePanel
-                              id={TOOLS_PANEL_ID}
-                              mode="collapsible"
-                              initialSize={50}
-                              minSize="25%"
-                              paddingSize="s"
-                              onToggleCollapsedInternal={() =>
-                                onToggleToolsChange()
-                              }
-                              style={{ marginBottom: '-16px' }}
-                            >
-                              <Tools
-                                workflow={workflow}
-                                ingestResponse={ingestResponse}
-                                queryResponse={queryResponse}
-                              />
-                            </EuiResizablePanel>
-                          </>
-                        );
-                      }}
-                    </EuiResizableContainer>
-                  </EuiResizablePanel>
-                </>
-              );
-            }}
-          </EuiResizableContainer>
-        </Form>
-      )}
-    </Formik>
+      {(EuiResizablePanel, EuiResizableButton, { togglePanel }) => {
+        if (togglePanel) {
+          collapseFnHorizontal.current = (panelId: string, { direction }) =>
+            togglePanel(panelId, { direction });
+        }
+        return (
+          <>
+            <EuiResizablePanel
+              id={WORKFLOW_INPUTS_PANEL_ID}
+              mode="main"
+              initialSize={60}
+              minSize="25%"
+              paddingSize="s"
+            >
+              <WorkflowInputs
+                workflow={props.workflow}
+                uiConfig={props.uiConfig}
+                setUiConfig={props.setUiConfig}
+                setIngestResponse={setIngestResponse}
+                setQueryResponse={setQueryResponse}
+                ingestDocs={props.ingestDocs}
+                setIngestDocs={props.setIngestDocs}
+                query={props.query}
+                setQuery={props.setQuery}
+              />
+            </EuiResizablePanel>
+            <EuiResizableButton />
+            <EuiResizablePanel
+              id={PREVIEW_PANEL_ID}
+              style={{
+                marginRight: isPreviewPanelOpen ? '-32px' : '0px',
+                marginBottom: isToolsPanelOpen ? '0px' : '24px',
+              }}
+              mode="collapsible"
+              initialSize={60}
+              minSize="25%"
+              paddingSize="s"
+              onToggleCollapsedInternal={() => onTogglePreviewChange()}
+            >
+              <EuiResizableContainer
+                className="workspace-panel"
+                direction="vertical"
+                style={{
+                  marginLeft: '-8px',
+                  marginTop: '-8px',
+                  padding: 'none',
+                }}
+              >
+                {(EuiResizablePanel, EuiResizableButton, { togglePanel }) => {
+                  if (togglePanel) {
+                    collapseFnVertical.current = (
+                      panelId: string,
+                      { direction }
+                    ) =>
+                      // ignore is added since docs are incorrectly missing "top" and "bottom"
+                      // as valid direction options for vertically-configured resizable panels.
+                      // @ts-ignore
+                      togglePanel(panelId, { direction });
+                  }
+                  return (
+                    <>
+                      <EuiResizablePanel
+                        mode="main"
+                        initialSize={60}
+                        minSize="25%"
+                        paddingSize="s"
+                        style={{ marginBottom: '-8px' }}
+                      >
+                        <EuiFlexGroup
+                          direction="column"
+                          gutterSize="s"
+                          style={{ height: '100%' }}
+                        >
+                          <EuiFlexItem>
+                            <Workspace
+                              workflow={props.workflow}
+                              uiConfig={props.uiConfig}
+                            />
+                          </EuiFlexItem>
+                        </EuiFlexGroup>
+                      </EuiResizablePanel>
+                      <EuiResizableButton />
+                      <EuiResizablePanel
+                        id={TOOLS_PANEL_ID}
+                        mode="collapsible"
+                        initialSize={50}
+                        minSize="25%"
+                        paddingSize="s"
+                        onToggleCollapsedInternal={() => onToggleToolsChange()}
+                        style={{ marginBottom: '-16px' }}
+                      >
+                        <Tools
+                          workflow={props.workflow}
+                          ingestResponse={ingestResponse}
+                          queryResponse={queryResponse}
+                        />
+                      </EuiResizablePanel>
+                    </>
+                  );
+                }}
+              </EuiResizableContainer>
+            </EuiResizablePanel>
+          </>
+        );
+      }}
+    </EuiResizableContainer>
   ) : (
     <EuiFlexGroup direction="column">
       <EuiFlexItem grow={3}>
